requirements.txt
####python 3.10
# absl-py==2.1.0
asgiref==3.8.1
# astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==3.2.7
# flatbuffers==25.2.10
# gast==0.6.0
# google-pasta==0.2.0
# grpcio==1.70.0
# h5py
# idna
# keras
# libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
# ml-dtypes==0.4.1
# mysql-connector-python==8.0.33
# mysqlclient==2.1.1
namex==0.0.8
# numpy
# opencv-python
# opt_einsum==3.4.0
# optree==0.14.0
packaging==24.2
protobuf==3.20.3
Pygments==2.19.1
pytz==2025.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
# tensorboard
# tensorboard-data-server
# tensorflow
# tensorflow-io-gcs-filesystem
# tensorflow_intel
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
python-dotenv